bridge-web: update wallet state on Tag
