+++ src/components/Status/index.ts
@@ -1,2 +1,4 @@
 export * from './Connected/Connected';
 export * from './Disconnected/Disconnected';
+export * from './NoAccounts/NoAccounts';
+export * from './NotInstalled/NoInstalled';
